Feature/import data & unit tests (#29)
This PR adds an import script that can be called externally by running:

`pipenv run invenio shell site/mex_invenio/scripts/import_data.py
<email> <filepath>`

it can handle test data and expects ndjson format.

Processing is done in parallel using multiprocessing to increase
efficiency. Estimated time to process ~50k records is less than half an
hour.

The module runs it's own logger and writes to the file
`import_data_log`. This configuration might not be optimal.

This PR also sets up unit tests. A section of the README guides you to
how to set up and run unit tests.

It includes a very basic unit test to ensure the landing page returns a
200 response and four unit tests that test the command line script to
import data.
 
This PR also adds a config variable to settings
`RECORD_METADATA_CREATOR` which is used to populate the creator of the
imported data. This needs to be a dict.

I am having an issue with importing all vocabulary fixtures for the unit
testing. This is currently not a problem.

---------

Signed-off-by: Hrafn Malmquist <hrafn.malmquist@gmail.com>
Co-authored-by: kush.varade.19 <kush.varade.19@gmail.com>