Feature/cf UI normalised vocab (#66)
This PR replaces #65 and does a couple of things.

It fixes #55, #62 and #40.

Custom fields are now sorted in alphabetical order, in order to better
track changes made to them. Custom fields now have contextual
information passed to the UI side as react `props`. This is defined in
the config, like so:

```python
{
'field': 'mex:creator',
'props': {
     'type': '/schema/entities/person#/identifier',
     'description': 'The author of the publication.'}
}
```

Both values are taken directly from the
[mex-model](https://github.com/robert-koch-institut/mex-model). The
default type value 'string' is not from the model.

The `type` is used to determine how to render the property in the
template.

The `description` is used as a tooltip accompanying the property header.

## Pref labels
The vocabularies in the mex-model are values by reference. In order to
extract these values, the mex-model is submoduled under the
`custom_fields` directory and the new script pref_labels can be run in
order to write to a jinja file that contains a dict that is imported in
the template and the appropriate value, based on the selected i8
language is displayed.

This is a clumsy way of incorporating the lookup. The rendering is being
called via
[invenio-communities](https://github.com/inveniosoftware/invenio-communities/blob/e136798e83a3e59378433daac7a45b151edbe930/invenio_communities/views/communities.py#L108)
so it is not easy to get at passing custom variables to the template.
The danger to writing directly to the jinja file is if it breaks and
then potentially the templating will break completely. Another (hacky)
approach is to add this dict as a config variable.

Note that this will need to be run as a separate cron job to the data
import.

## Pipfile update
The invenio-app-rdm is updated to v. 12.0.17.

---------

Signed-off-by: Hrafn Malmquist <hrafn.malmquist@gmail.com>
Co-authored-by: kush.varade.19 <kush.varade.19@gmail.com>