Hotfix/edtf field (#115)
This PR fixes #106

This will also fix a cause of #114 in that the error in indexing was
preventing a true comparison of records during import resulting in new
versions being created with identical content.

I have made a
[PR](https://github.com/inveniosoftware/invenio-records-resources/pull/635)
to the invenio-records-resources module, where the bug is. When that is
backported to the [5 maintenance
branch](https://github.com/inveniosoftware/invenio-records-resources/tree/main-5.x)
and a minor release made, this PR can be reverted.

---------

Co-authored-by: Steven Eardley <steve@eardley.xyz>
Co-authored-by: Aga Domanska <aga@cottagelabs.com>