remove file logging from import and s3 scripts - use redirection instead (#123)
Despite the image not changing, we were getting permission denied when
trying to write log files within `logs` dir in the pods. I've taken out
the file handler entirely - the logs in stdout is sufficient for
kubernetes, and if you want to log to file locally you can use output
redirection `>` if needed.